+++ .prettierrc.cjs
@@ -1,4 +1,4 @@
 // Copyright 2019-2021 @polkadot/extension authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export { canDerive } from './canDerive';
+module.exports = require('@polkadot/dev/config/prettier.cjs');